Revert "moving context implementation to sdk"
This reverts commit 4c2c4de6ff1ec9d6f8ef4b9663db4f4d27f07e01.
