+++ opentelemetry-sdk/src/opentelemetry/sdk/context/propagation/tracecontexthttptextformat.py
@@ -16,6 +16,7 @@ import re
 import typing
 
 import opentelemetry.trace as trace
+from opentelemetry.context import Context
 from opentelemetry.context.propagation import (
     Getter,
     HTTPExtractor,
@@ -24,8 +25,7 @@ from opentelemetry.context.propagation import (
     get_as_list,
     set_in_dict,
 )
-from opentelemetry.sdk.context import Context
-from opentelemetry.sdk.trace.propagation.context import (
+from opentelemetry.trace.propagation.context import (
     from_context,
     with_span_context,
 )
