+++ opentelemetry-sdk/tests/context/propagation/test_b3_format.py
@@ -17,7 +17,7 @@ import unittest
 import opentelemetry.sdk.context.propagation.b3_format as b3_format
 import opentelemetry.sdk.trace as trace
 import opentelemetry.trace as trace_api
-from opentelemetry.sdk.trace.propagation.context import from_context
+from opentelemetry.trace.propagation.context import from_context
 
 INJECTOR = b3_format.B3Injector
 EXTRACTOR = b3_format.B3Extractor
